refactor(ollama): split oversized setup.test.ts into focused test modules
